+++ E-Mart/src/java/Model/Smartphone.java
@@ -16,7 +16,7 @@ public class Smartphone extends BaseProduct {
     private Double screenSize;
     private String color;
 
-    public Smartphone(int productId, String productName, float price, int quantity, String photo1,
+    public Smartphone(int productId, String productName, int price, int quantity, String photo1,
             String brand, String modelName, String productDescription, String category,
             String storageCapacity, Double screenSize, String color) {
         super(productId, productName, price, quantity, photo1, brand, modelName, productDescription, category);
